// src/Module/Products/DTO/Product.DTO.ts
export class ProductDTO {
  id?: number;
  imgSmall?: string;
  imgBig?: string;
  nameProduct?: string;
  price?: number;
  color?: string;
  goodsInStock?: number;
  content?: string;
  type?: string;
  IdProduct?: number;
  manufacturer?: string;
  caliber?: number;
  velocity?: number;
  conditions?: string;
  ammo_type?: string;
  actions?: string;
  barrel_style?: string;
  fire_mode?: string;
  gun_weight?: number;
  loudness?: number;
  mechanism?: string;
  Ammo_Weight?: number;
  Pellet_Shape?: string;
  Pellet_Quantity?: string;
}